message-persistence-service/requirements.txt
kafka-python==2.0.1
loguru==0.5.3
requests==2.24.0
